feat: Added RAK prefix to some base classes
Most projects have their own base classes. Although these base classes can inherit from RAKBase.BaseXXX, when the base class of a project is in a framework and the main project references both RAKBase and this project, you must specify the ownership of BaseXXX. This is a hassle for the caller.

Most of the time, RAKBase.BaseXXX will only be used once, so renaming it would be a clearer choice.